Make ParensSeq public
