+++ gen/HaskellWorks/Data/BalancedParens/Gen.hs
@@ -14,19 +14,19 @@ module HaskellWorks.Data.BalancedParens.Gen
   ) where
 
 import Data.Coerce
-import Data.Semigroup                                      ((<>))
+import Data.Semigroup                             ((<>))
 import Data.Word
-import HaskellWorks.Data.BalancedParens.Internal.ParensSeq (ParensSeq)
+import HaskellWorks.Data.BalancedParens.ParensSeq (ParensSeq)
 import HaskellWorks.Data.Positioning
 import Hedgehog
 
-import qualified Data.Vector                                         as DV
-import qualified Data.Vector.Storable                                as DVS
-import qualified HaskellWorks.Data.BalancedParens.Internal.ParensSeq as PS
-import qualified HaskellWorks.Data.BalancedParens.RangeMinMax        as RMM
-import qualified HaskellWorks.Data.BalancedParens.RangeMinMax2       as RMM2
-import qualified Hedgehog.Gen                                        as G
-import qualified Hedgehog.Range                                      as R
+import qualified Data.Vector                                   as DV
+import qualified Data.Vector.Storable                          as DVS
+import qualified HaskellWorks.Data.BalancedParens.ParensSeq    as PS
+import qualified HaskellWorks.Data.BalancedParens.RangeMinMax  as RMM
+import qualified HaskellWorks.Data.BalancedParens.RangeMinMax2 as RMM2
+import qualified Hedgehog.Gen                                  as G
+import qualified Hedgehog.Range                                as R
 
 count :: MonadGen m => Range Count -> m Count
 count r = coerce <$> G.word64 (coerce <$> r)
