--- src/HaskellWorks/Data/BalancedParens/Internal/ParensSeq/Types.hs
@@ -1,5 +1,0 @@
-module HaskellWorks.Data.BalancedParens.Internal.ParensSeq.Types
-  ( ParensSeq
-  ) where
-
-import HaskellWorks.Data.BalancedParens.Internal.ParensSeq.Internal (ParensSeq)
