+++ src/HaskellWorks/Data/BalancedParens/ParensSeq/Types.hs
@@ -1,0 +1,5 @@
+module HaskellWorks.Data.BalancedParens.ParensSeq.Types
+  ( ParensSeq
+  ) where
+
+import HaskellWorks.Data.BalancedParens.Internal.ParensSeq (ParensSeq)
